delated comments, adet funcionality to prevent crashes.

+++ skrypt.py
@@ -17,31 +17,34 @@ def download_pdf(_url, title, cat):
     file_name = title
     print("Downloading book: {}".format(file_name))
     r = requests.get(_url[0], stream = True)
-    dirr = cat + "/" + title + ".pdf"
-    with open(dirr, "wb") as pdf:
-        for chunk in r.iter_content(chunk_size=10000000):
-            if chunk:
-                pdf.write(chunk)
+    dirr = cat + "/" + title + '.pdf'
+    #print (dirr)
+    if os.path.exists(dirr):
+        pass
+    else:
+        with open(dirr, "wb") as pdf:
+            for chunk in r.iter_content(chunk_size=10000000):
+                if chunk:
+                    pdf.write(chunk)
 
 if __name__ == '__main__':
-    path_ = os.getcwd() #mozna dac dir gdzie chcesz to zapisac, domyslnie tworzy foldery w katalogu ze skryptem
+    path_ = os.getcwd()
     xlsx_path = os.path.join(path_, 'm.xlsx')
-    df = pd.read_excel(xlsx_path) ## tu dajesz dir do xlsx'a
+    df = pd.read_excel(xlsx_path)
     n = df.shape[0]
 
 
     names = pd.unique(df['English Package Name'])
     for name in names:
-        name_ = os.path.join(path_, name)
-        os.mkdir(name_)
+        if not os.path.exists(name):
+            name_ = os.path.join(path_, name)
+            os.mkdir(name_)
 
     for i, _url in enumerate(df['OpenURL']):
-        title = df['Book Title'][i]
+        title = df['Book Title'][i].split('/')[0]
         category = df['English Package Name'][i]
         download_pdf(get_pdfs(_url),title, category)
         print("{}/{}".format(i+1,n))
-
-
 
 
 ###############################################
